Rework ownpod acquisition and decouple it from configmap writer
